4.18-4.23

+++ osa4/controllers/blogs.js
@@ -1,25 +1,27 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const userExtractor = require('../utils/userExtractor')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
         .find({})
-        .populate('user', {username:1, name:1, id:1})
+        .populate('user', { username: 1, name: 1, id: 1 })
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body
-    if (body.likes === undefined){
+
+    const user = await User.findById(request.user)
+    if (body.likes === undefined) {
         body.likes = 0
     }
-    if (body.title === undefined || body.url === undefined){
+    if (body.title === undefined || body.url === undefined) {
         response.status(400).send()
         return
     }
-
-    const user = await User.findById(body.userId)
 
     const blog = new Blog({
         _id: body._id,
@@ -30,19 +32,22 @@ blogsRouter.post('/', async (request, response) => {
         user: user._id,
         __v: body.__v
     })
-
     const posted = await blog.save()
     user.blogs = user.blogs.concat(posted._id)
     await user.save()
     response.status(201).json(posted)
 })
 
-blogsRouter.delete('/:id', async (request, response) =>{
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const blogToRemove = await Blog.findById(request.params.id)
+    if (blogToRemove.user.toString() !== request.user.toString()) {
+        return response.status(403).json({ error: "the blog to deleted wasn't created by the current user" })
+    }
     await Blog.findByIdAndRemove(request.params.id)
-    response.send(204).end()
+    response.status(200).send()
 })
 
-blogsRouter.put('/:id', async (request, response) =>{
+blogsRouter.put('/:id', async (request, response) => {
     const body = request.body
 
     const blog = {
@@ -54,7 +59,7 @@ blogsRouter.put('/:id', async (request, response) =>{
         __v: body.__v
     }
 
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
     response.json(updatedBlog)
 })
